prenotazione controller

+++ src/main/java/kevindonati/M5_W3_D5/services/PrenotazioneService.java
@@ -3,7 +3,9 @@ package kevindonati.M5_W3_D5.services;
 import kevindonati.M5_W3_D5.entities.Evento;
 import kevindonati.M5_W3_D5.entities.Prenotazione;
 import kevindonati.M5_W3_D5.entities.Utente;
+import kevindonati.M5_W3_D5.exceptions.BadRequestException;
 import kevindonati.M5_W3_D5.exceptions.NotFoundException;
+import kevindonati.M5_W3_D5.exceptions.UnauthorizedException;
 import kevindonati.M5_W3_D5.repositories.PrenotazioneRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -12,6 +14,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
 
@@ -23,6 +26,17 @@ public class PrenotazioneService {
     private EventoService eventoService;
     @Autowired
     private UtenteService utenteService;
+
+    public Prenotazione save(UUID idEvento, Utente utente) {
+        Evento eventoTrovato = eventoService.findById(idEvento);
+
+        if (eventoTrovato.getPostiDisponibili() <= 0) {
+            throw new BadRequestException("Non ci sono più posti disponibili");
+        }
+
+        Prenotazione nuovaPrenotazione = new Prenotazione(LocalDate.now(), utente, eventoTrovato);
+        return prenotazioneRepository.save(nuovaPrenotazione);
+    }
 
     public Page<Prenotazione> findAll(int page, int size, String orderBy) {
         if (size > 50) size = 50;
@@ -37,8 +51,11 @@ public class PrenotazioneService {
         return prenotazioneRepository.findById(id).orElseThrow(() -> new NotFoundException("Prenotazione con id " + id + " non trovata"));
     }
 
-    public void findByIdAndDelete(UUID id) {
+    public void findByIdAndDelete(UUID id, Utente utente) {
         Prenotazione prenotazioneTrovata = this.findById(id);
+        if (!prenotazioneTrovata.getUtente().getId().equals(utente.getId())) {
+            throw new UnauthorizedException("Non puoi eliminare una prenotazione che non è tua");
+        }
         prenotazioneRepository.delete(prenotazioneTrovata);
     }
